+++ include/hex.h
@@ -1,2 +1,4 @@
-unsigned char* hex2bytes(const char *, int *);
-char* bytes2hex(const unsigned char*, int);
+#include <stdlib.h>
+
+unsigned char* hex2bytes(const char *, size_t *);
+char* bytes2hex(const unsigned char*, size_t);
